Wrote API tests for comments, likes, liked

+++ backend/database/views.py
@@ -194,7 +194,7 @@ class PostsAPIs(viewsets.ViewSet):
         serializer = PostsSerializer(queryset, many=True)
         return Response(serializer.data)
 
-#completed
+#completed and tested
 class CommentsAPIs(viewsets.ViewSet):
 
     #TESTED
@@ -205,13 +205,13 @@ class CommentsAPIs(viewsets.ViewSet):
     def getComments(self, request, *args, **kwargs):
         postId = kwargs["postId"]
         if not Posts.objects.filter(id=postId).count() == 1:
-            return Response({"Tried to send non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"Tried to get comments from non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
 
         queryset = Comments.objects.filter(post_id=postId).order_by('-published')
         serializer = CommentsSerializer(queryset, many=True)
         return Response(serializer.data)
 
-    #NOT TESTED
+    #TESTED
     #
     #POST authors/{AUTHOR_ID}/posts/{POST_ID}/comments
     #if you post an object of “type”:”comment”, it will add your comment to the post whose id is POST_ID
@@ -222,9 +222,9 @@ class CommentsAPIs(viewsets.ViewSet):
 
         #check that authorId and postId exist
         if not Authors.objects.filter(id=authorId).count() ==1:
-            return Response({"Tried to send post to non-existent author"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"Tried to make comment from non-existent author"}, status=status.HTTP_400_BAD_REQUEST)
         if not Posts.objects.filter(id=postId).count() == 1:
-            return Response({"Tried to send non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
+            return Response({"Tried to make comment on non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
 
         author = Authors.objects.get(id=authorId)
         post = Posts.objects.get(id=postId)
@@ -233,13 +233,13 @@ class CommentsAPIs(viewsets.ViewSet):
         body = JSONParser().parse(io.BytesIO(request.body))
         contentType = 'text/plain'
         if 'contentType' in body:
-            if body['contentBody'] == 'text/markdown':
+            if body['contentType'] == 'text/markdown':
                 contentType = 'text/markdown'
-            elif body['contentBody'] != 'text/plain':
-                return Response({"Failed comment creation. Invalid input for contentBody."}, status=status.HTTP_400_BAD_REQUEST)
+            elif body['contentType'] != 'text/plain':
+                return Response({"Failed comment creation. Invalid input for contentType."}, status=status.HTTP_400_BAD_REQUEST)
 
-        if 'comment' in body:
-            comment = str(body['comment'])
+        if 'comment' in body and str(body['comment']).strip() != '':
+            comment = str(body['comment']).strip()
         else:
             return Response({"Failed comment creation. Missing 'comment' column."}, status=status.HTTP_400_BAD_REQUEST)
 
@@ -253,10 +253,10 @@ class CommentsAPIs(viewsets.ViewSet):
 
         return Response({"Comment Created Successfully"}, status=status.HTTP_200_OK)
 
-#completed
+#completed and tested
 class LikesAPIs(viewsets.ViewSet):
 
-    #NOT TESTED
+    #TESTED
     #
     #POST authors/{AUTHOR_ID}/posts/{POST_ID}/likes/{LIKER_ID}
     @action(detail=True, methods=['post'])
@@ -285,7 +285,7 @@ class LikesAPIs(viewsets.ViewSet):
 
         return Response("{Like created successfully}", status=status.HTTP_200_OK )
 
-    #NOT TESTED
+    #TESTED
     #
     #POST authors/{AUTHOR_ID}/posts/{POST_ID}/comments/{COMMENT_ID}/likes/{LIKER_ID}
     @action(detail=True, methods=['post'])
@@ -314,7 +314,7 @@ class LikesAPIs(viewsets.ViewSet):
 
         return Response("{Like created successfully}", status=status.HTTP_200_OK )
 
-    #NOT TESTED
+    #TESTED
     #
     #DELETE authors/{AUTHOR_ID}/posts/{POST_ID}/likes/{LIKER_ID}
     @action(detail=True, methods=['delete'])
@@ -329,12 +329,14 @@ class LikesAPIs(viewsets.ViewSet):
             return Response({"Tried to delete a like on a non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
         liker = Authors.objects.get(id=likerId)
         post = Posts.objects.get(id=postId)
-        
+
+        if Likes.objects.filter(post=post, author=liker).count() < 1:
+            return Response({"Tried to delete a non-existent like"}, status=status.HTTP_400_BAD_REQUEST)
         Likes.objects.filter(post=post, author=liker).delete()
         
         return Response({"Delete Like Successful"}, status=status.HTTP_200_OK)
 
-    #NOT TESTED
+    #TESTED
     #
     #DELETE authors/{AUTHOR_ID}/posts/{POST_ID}/comments/{COMMENT_ID}/likes/{LIKER_ID}
     @action(detail=True, methods=['delete'])
@@ -349,13 +351,15 @@ class LikesAPIs(viewsets.ViewSet):
             return Response({"Tried to delete a like on a non-existent comment"}, status=status.HTTP_400_BAD_REQUEST)
         liker = Authors.objects.get(id=likerId)
         comment = Comments.objects.get(id=commentId)
-        
-        Likes.objects.filter(comment=comment, author=liker).delete()
+
+        if LikesComments.objects.filter(comment=comment, author=liker).count() < 1:
+            return Response({"Tried to delete a non-existent like"}, status=status.HTTP_400_BAD_REQUEST)
+        LikesComments.objects.filter(comment=comment, author=liker).delete()
         
         return Response({"Delete Like Successful"}, status=status.HTTP_200_OK)
 
 
-    #NOT TESTED
+    #TESTED
     #
     #GET authors/{AUTHOR_ID}/posts/{POST_ID}/likes/inbox?page=value&size=value
     #a list of likes from other authors on AUTHOR_ID’s post POST_ID
@@ -367,6 +371,9 @@ class LikesAPIs(viewsets.ViewSet):
             size = int(request.GET.get('size',10))
         except:
             return Response("{Page or Size not an integer}", status=status.HTTP_400_BAD_REQUEST )
+
+        if Posts.objects.filter(id=postId).count() < 1:
+            return Response({"Tried to get likes from a non-existent post"}, status=status.HTTP_400_BAD_REQUEST)
             
         likeObjs = []
         for like in Likes.objects.filter(post_id=postId):
@@ -381,7 +388,7 @@ class LikesAPIs(viewsets.ViewSet):
         
         return Response(output, status=status.HTTP_200_OK)
 
-    #NOT TESTED
+    #TESTED
     #
     #GET authors/{AUTHOR_ID}/posts/{POST_ID}/comments/{COMMENT_ID}/likes/inbox?page=value&size=value
     #a list of likes from other authors on AUTHOR_ID’s post POST_ID comment COMMENT_ID
@@ -393,6 +400,9 @@ class LikesAPIs(viewsets.ViewSet):
             size = int(request.GET.get('size',10))
         except:
             return Response("{Page or Size not an integer}", status=status.HTTP_400_BAD_REQUEST )
+
+        if Comments.objects.filter(id=commentId).count() < 1:
+            return Response({"Tried to get likes from a non-existent comment"}, status=status.HTTP_400_BAD_REQUEST)
 
         commentLikeObjs = []
         for like in LikesComments.objects.filter(comment_id=commentId):
@@ -409,7 +419,7 @@ class LikesAPIs(viewsets.ViewSet):
 
 #completed and tested
 class LikedAPIs(viewsets.ViewSet):
-    #NOT TESTED
+    #TESTED
     #
     #GET authors/{AUTHOR_ID}/posts/{POST_ID}/like/{LIKER_ID}
     #returns true if authorId has made a like on postId, otherwise false
@@ -431,7 +441,7 @@ class LikedAPIs(viewsets.ViewSet):
             return Response(True, status=status.HTTP_200_OK)
         return Response(False, status=status.HTTP_200_OK)
 
-    #NOT TESTED
+    #TESTED
     #
     #GET authors/{AUTHOR_ID}/liked/inbox?page=value&size=value
     #list what public things AUTHOR_ID liked
@@ -443,6 +453,9 @@ class LikedAPIs(viewsets.ViewSet):
             size = int(request.GET.get('size',10))
         except:
             return Response("{Page or Size not an integer}", status=status.HTTP_400_BAD_REQUEST )
+
+        if not Authors.objects.filter(id=authorId).count() == 1:
+            return Response({"Tried to check likes of a non-existent author"}, status=status.HTTP_400_BAD_REQUEST)
 
         likeObjs = []
         for like in Likes.objects.filter(author_id=authorId):
